requirements.txt
opencv-python
torch
torchvision
ultralytics
requests
python-dotenv
